Rebase RDP-RAIL support from microsoft's mirror
I call this a rebase, but it also includes fixes to make it build,
and a general merge of many parts trying to reduce the ammout of
duplicated code.

- Includes everything since the original commit:
"Merged PR 4742168: Add preliminary RDP-RAIL support for weston"
https://github.com/microsoft/weston-mirror/commit/f590a956c378eb2ab765d05cb2331386f19f8704

- I have merged their 'rdprail-shell' folder with the 'desktop-shell'
  because they basically copy/pasted 'desktop-shell' and modified it
slightly to make 'rdprail-shell'. This reduces duplicated code by so
much that I think it's worth it.

- the "wslgd-notify" plugin isn't included. This looks VERY ms-specific,
so I'm unsure it's worth it including it. Could be wrong of course.

- (I tried to move the backend-specific parts of the rdp-rail
  implementation in a desktop-shell/rdprail/backend.c but for the love
of god I can't figure how to make it build without recursion loops...)

- Current state: it builds, but crashes (ofc). This will take.. time.

Co-Authored-By: gpotter2 <10530980+gpotter2@users.noreply.github.com>